pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "indic-tokenization-lab"
version = "0.1.0"
description = "Tokenizer stack for Devanagari (Hindi + Sanskrit) with morphology-aware segmentation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Indic Tokenization Lab Contributors"}
]
keywords = ["tokenization", "nlp", "devanagari", "hindi", "sanskrit", "indic", "bpe", "morphology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "transformers>=4.20.0",
    "tokenizers>=0.13.0",
    "sentencepiece>=0.1.96",
    "pyyaml>=6.0",
    "regex>=2023.0.0",
    "tiktoken>=0.5.0; extra == 'openai'",
]

[project.optional-dependencies]
openai = ["tiktoken>=0.5.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
indic-compare = "scripts.compare_tokenizers:main"
indic-benchmark = "scripts.run_benchmark:main"

[tool.setuptools]
packages = ["tokenizers", "eval", "scripts"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py38"

